[#15] improving second test with dpckan_test module
